evldns_udp_read_callback(): Use a malloc'd buffer for the wire request
This commit converts evldns_udp_read_callback() to use a malloc'd buffer
for the incoming wire request rather than one allocated on the stack,
and stores a reference to this buffer in the corresponding
'evldns_server_request' object's .wire_request field. The number of
bytes received from the socket is stored in the .wire_reqlen field.

This makes the original request as read from the wire available for UDP
requests, as is currently done for TCP requests.

(Fixes #4.)