Allow dropping requests if the callback chain doesn't want to answer
This commit allows the callback chain to drop a request on the floor by
setting the 'response' field to NULL. This is less polite than returning
REFUSED or another rcode but may be desireable in certain situations.

This depends on the previous commit ("Cleanup the TCP connection if
server_process_packet() fails"). Without that fix, the server goes into
an infinite loop if the callback chain drops a query that arrived via
TCP.