Support blackholing of requests
This patch adds a 'blackhole' flag to evldns_server_request. If a
callback in the callback chain sets this flag, processing of the
callback chain will be aborted and the request will be "blackholed".

A blackholed UDP request is dropped on the floor by the server.  No
response is sent to the client.

A blackholed TCP connection is immediately closed by the server.